requirements.txt
-i https://pypi.org/simple
amqp==5.1.1 ; python_version >= '3.6'
asgiref==3.7.2 ; python_version >= '3.7'
async-timeout==4.0.3 ; python_full_version <= '3.11.2'
babel==2.12.1
billiard==4.1.0 ; python_version >= '3.7'
boto3==1.28.27
botocore==1.31.27 ; python_version >= '3.7'
celery==5.3.1
certifi==2023.7.22 ; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==3.2.0 ; python_full_version >= '3.7.0'
click==8.1.6 ; python_version >= '3.7'
click-didyoumean==0.3.0 ; python_full_version >= '3.6.2' and python_full_version < '4.0.0'
click-plugins==1.1.1
click-repl==0.3.0 ; python_version >= '3.6'
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.3 ; python_version >= '3.7'
defusedxml==0.7.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django==4.2.4
django-cors-headers==4.2.0
django-filter==23.2
django-pandas==0.6.6
django-phonenumber-field[phonenumbers]==7.1.0
django-redis==5.3.0
django-rest-swagger==2.2.0
django-storages==1.13.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.4 ; python_version >= '3.5'
inflection==0.5.1 ; python_version >= '3.5'
itypes==1.2.0
jinja2==3.1.2 ; python_version >= '3.7'
jmespath==1.0.1 ; python_version >= '3.7'
kombu==5.3.1 ; python_version >= '3.8'
markupsafe==2.1.3 ; python_version >= '3.7'
numpy==1.25.2 ; python_version < '3.10'
oauthlib==3.2.2 ; python_version >= '3.6'
openapi-codec==1.3.2
packaging==23.1 ; python_version >= '3.7'
pandas==2.0.3
phonenumbers==8.13.18
pillow==10.0.0
prompt-toolkit==3.0.39 ; python_full_version >= '3.7.0'
psycopg2==2.9.7
pycparser==2.21
pyjwt==2.8.0 ; python_version >= '3.7'
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
pyyaml==6.0.1 ; python_version >= '3.6'
redis==5.0.0
requests==2.31.0 ; python_version >= '3.7'
requests-oauthlib==1.3.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
s3transfer==0.6.2 ; python_version >= '3.7'
simplejson==3.19.1 ; python_version >= '2.5' and python_version not in '3.0, 3.1, 3.2, 3.3'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
social-auth-app-django==5.2.0
social-auth-core==4.4.2 ; python_version >= '3.6'
sqlparse==0.4.4 ; python_version >= '3.5'
typing-extensions==4.7.1 ; python_version < '3.11'
tzdata==2023.3 ; python_version >= '2'
uritemplate==4.1.1 ; python_version >= '3.6'
urllib3==1.26.16 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
vine==5.0.0 ; python_version >= '3.6'
wcwidth==0.2.6
whitenoise==6.5.0
